+++ src/helpers.js
@@ -24,6 +24,7 @@ export const defaultOptions = {
   cleanCacheThrottle: 0,
 };
 
+export const SELECTORS_TAG = "selector";
 export const INIT = "init";
 export const CLEAN_CACHE = "cleanCache";
 export const RESET_STATE = "resetState";