+++ test/providers/config.js
@@ -11,7 +11,7 @@ Unless required by applicable law or agreed to in writing, software distributed 
 
 const sinon = require("sinon");
 
-const { Provider, providers } = require("../../src/index");
+const { Provider, Selector, providers } = require("../../src/index");
 const { defaultOptions } = require("../../src/helpers");
 
 describe("providers handler config method", () => {
@@ -331,6 +331,66 @@ describe("providers handler config method", () => {
         });
 
         expect(fooProvider6.tags).toEqual(["tag-3", "tag-5", "tag-6"]);
+      });
+    });
+  });
+
+  describe("selectors tag", () => {
+    let fooSelector;
+    let fooSelector2;
+    beforeEach(() => {
+      fooSelector = new Selector(() => {});
+      fooSelector2 = new Selector(() => {}, { tags: "tag-5" });
+    });
+
+    describe("tags getter", () => {
+      it("should return selector tag extended with tags option", () => {
+        expect(fooSelector.tags).toEqual(["selector"]);
+        expect(fooSelector2.tags).toEqual(["selector", "tag-5"]);
+      });
+    });
+
+    describe("when applied to groups of providers using getByTag method", () => {
+      it("should apply config to all providers in group", () => {
+        providers.getByTag("selector").config({
+          foo: "foo",
+        });
+
+        expect(fooSelector.options.foo).toEqual("foo");
+        expect(fooSelector2.options.foo).toEqual("foo");
+      });
+
+      it("should not apply config to non-belonging selectors", () => {
+        providers.getByTag("tag-5").config({
+          foo: "foo",
+        });
+
+        expect(fooSelector2.options.foo).toEqual("foo");
+        expect(fooSelector.options.foo).toEqual(undefined);
+      });
+
+      it("should extend previously defined configuration when creating source containing tag", () => {
+        providers.getByTag("selector").config({
+          foo: "foo",
+          foo2: "foo2",
+        });
+
+        const fooSelector3 = new Selector(() => {}, { tags: ["tag-5", "tag-6"] });
+
+        providers.getByTag("tag-6").config({
+          foo2: "new-foo2",
+          foo3: "foo3",
+        });
+
+        expect(fooSelector3.options).toEqual({
+          ...defaultOptions,
+          foo: "foo",
+          foo2: "new-foo2",
+          foo3: "foo3",
+          tags: ["tag-5", "tag-6"],
+        });
+
+        expect(fooSelector3.tags).toEqual(["selector", "tag-5", "tag-6"]);
       });
     });
   });