+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 ### Added
 - feat: Allow defining baseTags property in addons. Will be extended with "tags" option when instantiated for calculating provider tags. (#143)
-- feat: Add tags getter.
+- feat: Add tags getter
+- feat: Add "selector" base tag to selectors (#163)
 ### Changed
 ### Fixed
 - docs(readme): Fix typos